block: Move DiskTopology to common code
It doens't make sense to limit this to the async_io module and import
across modules so place in default create module.
